renamed login folder to login-page, added route to logo
