+++ src/app/core/auth/auth.module.ts
@@ -1,15 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { LoginComponent } from './login/login.component';
+import { LoginComponent } from './login-page/login.component';
 import { ReactiveFormsModule } from '@angular/forms';
 
 @NgModule({
-  declarations: [
-    LoginComponent
-  ],
-  imports: [
-    CommonModule, 
-    ReactiveFormsModule
-  ]
+  declarations: [LoginComponent],
+  imports: [CommonModule, ReactiveFormsModule],
 })
-export class AuthModule { }
+export class AuthModule {}
