+++ src/app/core/components/header/header.component.html
@@ -1,5 +1,5 @@
 <header>
-  <h1>Fake Cafe!</h1>
+  <h1 routerLink="/">Fake Cafe!</h1>
   <a routerLink="/login">
     <button mat-raised-button class="login_box" >
       <mat-icon fontIcon="login" class="login"></mat-icon>
