+++ src/app/core/core-routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { LoginComponent } from './auth/login/login.component';
+import { LoginComponent } from './auth/login-page/login.component';
 
-const routes: Routes = [{path: 'login', component: LoginComponent}];
+const routes: Routes = [{ path: 'login', component: LoginComponent }];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
